# monolithic/backend/app/admin/service/cmd/server/assets/openapi.yaml
...
openapi: 3.0.3
info:
    title: Kratos巨石应用实践API
    description: Kratos巨石应用实践API
    contact:
        name: tx7do
        url: https://github.com/tx7do/kratos-monolithic-demo
        email: yanglinbo@gmail.com
    license:
        name: MIT License
        url: https://github.com/tx7do/kratos-monolithic-demo/blob/master/LICENSE
    version: "1.0"
paths:
...
    /admin/v1/menus:
        get:
            tags:
                - MenuService
            description: 查询菜单列表
            operationId: MenuService_ListMenu
            parameters:
                - name: page
                  in: query
                  description: 当前页码
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  description: 每页的行数
                  schema:
                    type: integer
                    format: int32
                - name: query
                  in: query
                  description: 与过滤参数
                  schema:
                    type: string
                - name: or
                  in: query
                  description: 或过滤参数
                  schema:
                    type: string
                - name: orderBy
                  in: query
                  description: 排序条件
                  schema:
                    type: array
                    items:
                        type: string
                - name: nopaging
                  in: query
                  description: 是否不分页
                  schema:
                    type: boolean
                - name: fieldMask
                  in: query
                  description: 字段掩码
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListMenuResponse'
            security:
                - OAuth2PasswordBearer: []
        post:
            tags:
                - MenuService
            description: 创建菜单
            operationId: MenuService_CreateMenu
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateMenuRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
            security:
                - OAuth2PasswordBearer: []
    /admin/v1/menus/{id}:
        get:
            tags:
                - MenuService
            description: 查询菜单详情
            operationId: MenuService_GetMenu
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Menu'
            security:
                - OAuth2PasswordBearer: []
        delete:
            tags:
                - MenuService
            description: 删除菜单
            operationId: MenuService_DeleteMenu
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: int32
                - name: operatorId
                  in: query
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content: {}
            security:
                - OAuth2PasswordBearer: []
    /admin/v1/menus/{menu.id}:
        put:
            tags:
                - MenuService
            description: 更新菜单
            operationId: MenuService_UpdateMenu
            parameters:
                - name: menu.id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: operatorId
                  in: query
                  schema:
                    type: integer
                    format: uint32
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Menu'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
            security:
                - OAuth2PasswordBearer: []
...
components:
    schemas:
...
        CreateMenuRequest:
            type: object
            properties:
                menu:
                    $ref: '#/components/schemas/Menu'
                operatorId:
                    type: integer
                    format: uint32
            description: 创建菜单 - 请求
...
        ListMenuResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/Menu'
                total:
                    type: integer
                    format: int32
            description: 查询菜单列表 - 回应
...
        Menu:
            type: object
            properties:
                id:
                    type: integer
                    format: int32
                parentId:
                    type: integer
                    format: int32
                orderNo:
                    type: integer
                    format: int32
                name:
                    type: string
                type:
                    enum:
                        - FOLDER
                        - MENU
                        - BUTTON
                    type: string
                    default: FOLDER
                    description: 类型
                    format: enum
                status:
                    enum:
                        - ON
                        - OFF
                    type: string
                    default: ON
                    description: 状态
                path:
                    type: string
                title:
                    type: string
                component:
                    type: string
                icon:
                    type: string
                permissions:
                    type: array
                    items:
                        type: string
                children:
                    type: array
                    items:
                        $ref: '#/components/schemas/Menu'
                keepAlive:
                    type: boolean
                show:
                    type: boolean
                isExt:
                    type: boolean
                extUrl:
                    type: string
                hideBreadcrumb:
                    type: boolean
                hideTab:
                    type: boolean
                hideMenu:
                    type: boolean
                hideChildrenInMenu:
                    type: boolean
                hidePathForChildren:
                    type: boolean
                currentActiveMenu:
                    type: string
                redirect:
                    type: string
                createTime:
                    type: string
                updateTime:
                    type: string
                deleteTime:
                    type: string
            description: 菜单
...
    securitySchemes:
        OAuth2PasswordBearer:
            type: oauth2
            flows:
                password:
                    tokenUrl: /admin/v1/login
                    refreshUrl: /admin/v1/refresh_token
                    scopes: {}
...